requirements.txt
# Основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# Whisper (CPU версия)
openai-whisper==20231117

# Для обработки аудио
pydub==0.25.1
librosa==0.10.1

# Дополнительные
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psutil==5.9.6

# Для аналитики/визуализации
pandas==2.1.4
numpy==1.26.2

# Для HTTP запросов
httpx==0.25.2
requests==2.31.0